+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - add support in add command for multiple repos
 - Add constants.dart and change master folder to .master
 - Add tests for creation of .organizations
+- Add test: logs error when primary and all fallback organization clones fail
 
 ### Changed